requirements.txt
bleach==2.1.2
diff-match-patch==20121119
Django==1.11.6
django-crispy-forms==1.7.0
django-formtools==2.1
django-haystack==2.7.0
django-import-export==1.0.0
django-pure-pagination==0.3.0
django-reversion==2.0.13
et-xmlfile==1.0.1
future==0.16.0
html5lib==1.0.1
httplib2==0.9.2
jdcal==1.3
jieba==0.39
Markdown==2.6.11
odfpy==1.3.6
openpyxl==2.5.0
Pillow==5.0.0
Pygments==2.2.0
PyMySQL==0.8.0
pytz==2018.3
PyYAML==3.12
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
webencodings==0.5.1
Whoosh==2.7.4
xlrd==1.1.0
xlwt==1.3.0
